Moved OrderHistory.java to src/main
I also made several changes to the original file in order to be able to read and append the buyer events into the order history CSV file. Removed the customer ID variable in hopes that the email variable should be able to replicate any functionality the customer ID variable would have been able to do.